add method in DimensionDtoImpl.class

+++ src/main/java/com/CezaryZal/dto/DimensionDTO.java
@@ -2,7 +2,7 @@ package com.CezaryZal.dto;
 
 public interface DimensionDTO {
 
-double getValueOfDimension();
+int getValueOfDimension();
 
 double getLowerDeviation();
